fix(point): Make point.pattern legend available for any chart type
Custom point only was available on axis based charts.
This fix will make available for all charts.

Ref #3653
